+++ src/components/TheHeaderProgress.vue
@@ -1,0 +1,17 @@
+<template>
+    <a href="#" class="text-sm">
+      <div v-if="false" class="flex items-center justify-between gap-1">
+        Day complete !
+        <CheckCircleIcon class="w-7 text-green-500" />
+      </div>
+      <div v-else class="flex items-center justify-between gap-1">
+        Progress: <span class="font-mono">20%</span>
+        <div class="h-3 w-3 rounded-full bg-red-500"></div>
+      </div>
+    </a>
+</template>
+
+<script setup>
+  import { CheckCircleIcon } from '@heroicons/vue/24/solid';
+
+</script>
