+++ src/components/TheLogo.vue
@@ -1,0 +1,5 @@
+<template>
+  <a href="#"><img src="../assets/images/logo.png" alt="Logo" class="h-full w-48" /></a>
+</template>
+
+<script setup></script>
